Make the scaling factor be a double so that it can be arbitrarily scaled; add automatic scaling; many scaling tweaks. Now works up to 100x100 (though slow)

+++ BlackSymbol.java
@@ -16,7 +16,7 @@ public class BlackSymbol extends Actor
     {
         // Add your action code here.
     }
-    public BlackSymbol(int scaleFactor) {
-        getImage().scale(60/scaleFactor,60/scaleFactor);
+    public BlackSymbol(double scaleFactor) {
+        getImage().scale((int)(60/scaleFactor),(int)(60/scaleFactor));
     }
 }
